fix SonarCube bugs

+++ key/key.core/src/main/java/de/uka/ilkd/key/nparser/builder/ExpressionBuilder.java
@@ -1671,7 +1671,7 @@ public class ExpressionBuilder extends DefaultBuilder {
         }
 
         KeYJavaType kjt = null;
-        final int packageEnd = i - 1;
+        final long packageEnd = i - 1;
         for (i = 0; i < parts.size(); i++) {
             String test = parts.stream()
                     //.skip(packageEnd)
